requirements.txt
feedparser
requests
trafilatura
gspread
google-auth
python-dateutil
beautifulsoup4
lxml
requests
python-dotenv